detangle tests ci fixes
