+++ curves/bls12_381/benches/bls12_381.rs
@@ -1,8 +1,8 @@
 use ark_algebra_bench_templates::*;
-use ark_ec_bench_templates::*;
 use ark_bls12_381::{
     fq::Fq, fq2::Fq2, fr::Fr, Bls12_381, Fq12, G1Projective as G1, G2Projective as G2,
 };
+use ark_ec_bench_templates::*;
 
 bench!(
     Name = "Bls12_381",
