+++ curves/bn254/benches/bn254.rs
@@ -1,6 +1,6 @@
 use ark_algebra_bench_templates::*;
-use ark_ec_bench_templates::*;
 use ark_bn254::{fq::Fq, fq2::Fq2, fr::Fr, Bn254, Fq12, G1Projective as G1, G2Projective as G2};
+use ark_ec_bench_templates::*;
 
 bench!(
     Name = "BN254",
