+++ test-curves/Cargo.toml
@@ -22,6 +22,7 @@ workspace = true
 ark-std = { workspace = true, default-features = false }
 ark-ff = { workspace = true, default-features = false }
 ark-ec = { workspace = true, default-features = false }
+ark-test-fields = { workspace = true, default-features = false }
 
 [dev-dependencies]
 ark-serialize = { workspace = true, default-features = false }
