+++ test-curves/src/bn384_small_two_adicity/mod.rs
@@ -1,12 +1,8 @@
 #[cfg(feature = "bn384_small_two_adicity_base_field")]
-pub mod fq;
-#[cfg(feature = "bn384_small_two_adicity_base_field")]
-pub use fq::*;
+pub use ark_test_fields::bn384_small_two_adicity::fq::*;
 
 #[cfg(feature = "bn384_small_two_adicity_scalar_field")]
-pub mod fr;
-#[cfg(feature = "bn384_small_two_adicity_scalar_field")]
-pub use fr::*;
+pub use ark_test_fields::bn384_small_two_adicity::fr::*;
 
 #[cfg(feature = "bn384_small_two_adicity_curve")]
 pub mod g1;
