+++ test-curves/src/mnt6_753/mod.rs
@@ -1,11 +1,4 @@
-pub mod fr;
-pub use self::fr::*;
-
-pub mod fq;
-pub use self::fq::*;
-
-pub mod fq3;
-pub use self::fq3::*;
+pub use ark_test_fields::mnt6_753::*;
 
 #[cfg(test)]
 mod tests;
