+++ test-fields/src/bls12_381/fr.rs
@@ -7,3 +7,21 @@ use ark_ff::fields::{Fp256, MontBackend, MontConfig};
 #[small_subgroup_power = "1"]
 pub struct FrConfig;
 pub type Fr = Fp256<MontBackend<FrConfig, 4>>;
+
+#[test]
+fn test_inv() {
+    assert_eq!(FrConfig::INV, 0xffff_fffe_ffff_ffff);
+}
+
+#[test]
+fn test_modulus() {
+    assert_eq!(
+        FrConfig::MODULUS.0,
+        [
+            0xffff_ffff_0000_0001,
+            0x53bd_a402_fffe_5bfe,
+            0x3339_d808_09a1_d805,
+            0x73ed_a753_299d_7d48,
+        ]
+    );
+}
