+++ test-fields/src/bls12_381/mod.rs
@@ -1,0 +1,8 @@
+pub mod fr;
+pub use fr::*;
+
+pub mod fq;
+pub mod fq12;
+pub mod fq2;
+pub mod fq6;
+pub use {fq::*, fq12::*, fq2::*, fq6::*};
