+++ test-fields/src/bn384_small_two_adicity/mod.rs
@@ -1,0 +1,5 @@
+pub mod fq;
+pub use fq::*;
+
+pub mod fr;
+pub use fr::*;
