+++ test-fields/src/ed_on_bls12_381/mod.rs
@@ -1,0 +1,5 @@
+pub mod fq;
+pub use fq::*;
+
+pub mod fr;
+pub use fr::*;
