+++ test-fields/src/lib.rs
@@ -5,7 +5,9 @@ pub use ark_ff::{self, fields::models::*, FftField, Field, LegendreSymbol, MontF
 pub mod fp128;
 pub mod smallfp;
 
-pub mod bls12_381 {
-    mod fr;
-    pub use fr::*;
-}
+pub mod bls12_381;
+pub mod bn384_small_two_adicity;
+pub mod ed_on_bls12_381;
+pub mod mnt4_753;
+pub mod mnt6_753;
+pub mod secp256k1;
