+++ test-fields/src/mnt4_753/mod.rs
@@ -1,0 +1,3 @@
+pub mod fq;
+pub mod fr;
+pub use {fq::*, fr::*};
