+++ test-fields/src/secp256k1/mod.rs
@@ -1,0 +1,5 @@
+pub mod fq;
+pub use fq::*;
+
+pub mod fr;
+pub use fr::*;
